--- src/Tests/Tests/IISTests.cs
@@ -1,27 +1,0 @@
-﻿#region Using Statements
-    using System;
-    using System.IO;
-    using System.Collections.ObjectModel;
-
-    using Xunit;
-
-    using Microsoft.Web.Administration;
-
-    using Cake.Core.Diagnostics;
-    using Cake.Core.IO;
-    using Cake.IIS;
-#endregion
-
-
-
-namespace Cake.IIS.Tests
-{
-    public class IISTests
-    {
-        [Fact]
-        public void Placeholder()
-        {
-
-        }
-    }
-}
